Fix some crashes with wide strings StringFormat on Linux.
- GCC template instantation doesn't work the same as MSVC's, it takes the
first one that fits, so our basic_string specialization of StringFormatArg
was never read by GCC, because it was declared afterwards, so in the case of passing
std::strings to StringFormat, it didn't redirect to its internal buffer.

- Another crash was due to the fact that std::swprintf doesn't work the same on Linux
than on Windows. Basically on Linux if the buffer doesn't fit it never returns the size
it would have to be. So we have to change the implementation to basically increment the
buffer size until it fits (currently I've chosen a 64 bytes step).